+++ pkg/commands/pull_request_test.go
@@ -144,7 +144,7 @@ func TestCreatePullRequest(t *testing.T) {
 
 	for _, s := range scenarios {
 		t.Run(s.testName, func(t *testing.T) {
-			gitCommand := newDummyGitCommand()
+			gitCommand := NewDummyGitCommand()
 			gitCommand.OSCommand.command = s.command
 			gitCommand.OSCommand.Config.GetUserConfig().Set("os.openLinkCommand", "open {{link}}")
 			dummyPullRequest := NewPullRequest(gitCommand)
